requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.12.5
pydantic-settings==2.12.0
httpx

# Application Dependencies
networkx==3.1
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
python-multipart==0.0.20

# Cryptography for Federated Blockchain
cryptography==41.0.7

# Machine Learning & AI Detection
transformers
torch
sentencepiece
peft

# Azure AI Services (Microsoft Imagine Cup 2026)
openai>=1.0.0  # Azure OpenAI SDK
azure-ai-contentsafety>=1.0.0  # Azure Content Safety SDK
azure-core>=1.28.0  # Azure core dependencies

# Ollama Python Client (deprecated, keeping for backward compatibility)
ollama==0.4.2

# Testing
pytest==7.4.4

# Server-Sent Events for real-time updates
sse-starlette==1.6.5

# For auto commit
GitPython
watchdog